--- casper/integer/integer_protocol.py
@@ -1,10 +1,0 @@
-from casper.integer.integer_view import IntegerView
-from casper.integer.bet import Bet
-from casper.integer.integer_plot_tool import IntegerPlotTool
-from casper.protocol import Protocol
-
-
-class IntegerProtocol(Protocol):
-    View = IntegerView
-    Message = Bet
-    PlotTool = IntegerPlotTool
