+++ casper/protocols/binary/binary_view.py
@@ -1,6 +1,6 @@
 """The binary view module extends a view for binary data structures """
-from casper.binary.bet import Bet
-from casper.integer.integer_view import IntegerView
+from casper.protocols.binary.bet import Bet
+from casper.protocols.integer.integer_view import IntegerView
 
 
 class BinaryView(IntegerView):
