+++ simulations/utils.py
@@ -2,9 +2,9 @@
 import itertools
 import random as r
 
-from casper.blockchain.blockchain_protocol import BlockchainProtocol
-from casper.binary.binary_protocol import BinaryProtocol
-from casper.integer.integer_protocol import IntegerProtocol
+from casper.protocols.blockchain.blockchain_protocol import BlockchainProtocol
+from casper.protocols.binary.binary_protocol import BinaryProtocol
+from casper.protocols.integer.integer_protocol import IntegerProtocol
 
 from casper.validator_set import ValidatorSet
