+++ tests/casper/test_protocol.py
@@ -1,9 +1,9 @@
 import pytest
 
 from casper.protocol import Protocol
-from casper.blockchain.blockchain_protocol import BlockchainProtocol
-from casper.binary.binary_protocol import BinaryProtocol
-from casper.integer.integer_protocol import IntegerProtocol
+from casper.protocols.blockchain.blockchain_protocol import BlockchainProtocol
+from casper.protocols.binary.binary_protocol import BinaryProtocol
+from casper.protocols.integer.integer_protocol import IntegerProtocol
 
 
 @pytest.mark.parametrize(
